fix: Cast fileBytes.length to Long for setFileSize
Fixed compilation error where int cannot be converted to Long.
Changed fileBytes.length to (long) fileBytes.length in site survey upload.

+++ src/main/java/com/magictech/modules/projects/ProjectsStorageController.java
@@ -2760,7 +2760,7 @@ public class ProjectsStorageController extends BaseModuleController {
                 surveyData.setWorkflowId(request.getId()); // Using request ID as workflow ID
                 surveyData.setExcelFile(fileBytes);
                 surveyData.setFileName(selectedFile.getName());
-                surveyData.setFileSize(fileBytes.length);
+                surveyData.setFileSize((long) fileBytes.length);
                 surveyData.setMimeType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet");
                 surveyData.setParsedData(parsedData);
                 surveyData.setSurveyDoneBy("PROJECT");
